evm file to runtime file
